+++ public/login.html
@@ -32,7 +32,7 @@
                       Looks good!
                     </div>
                   </div>
-                  <a href="./forgotPassword.html" class="text-danger mb-5">Forgot Password ?</a>
+                  <!-- <a href="./forgotPassword.html" class="text-danger mb-5">Forgot Password ?</a> -->
 
                   <div class="d-grid col-12 mx-auto">
